Modify methods

+++ src/main/java/org/wingate/fontbox/table/TableIO.java
@@ -3,5 +3,5 @@ package org.wingate.fontbox.table;
 import java.nio.ByteBuffer;
 
 public interface TableIO {
-    public void read(ByteBuffer buffer, long offset, long length);
+    public void read(ByteBuffer buffer, int offset, int length);
 }
